Use `outcome` field for checking CircleCI builds.
When checking to see if a CircleCI build passed, we were previously
checking the `status` field on Circle's webhook payload. This field
works for normal successful builds. However, if the build was previously
failing, and the build we are being notified about succeeds the `status`
field is actually set to the value `fixed` instead of `success`.

This change looks at the `outcome` field instead of the `success` field,
which is always `success` for successful builds.