Return all ECR images, not just the first page
Unfortunately, ECR images do not have a guaranteed sort order and we've
seen the latest image actually appear on the _second_ page of the
DescribeImages response. For now, collect _all_ images for a repository
and rely on limiting the number of available images at the ECR side.

As we're now potentially loading huge numbers of images, use destructive
array operations to convert them into JSON and sort them.