Only set Docker image if it isn't supplied
This change only explicitly overrides the Docker image parameter of a
task if it isn't already set. Rocksteady uses this method to decide
which docker image to deploy at any time, but this requires that all
tasks use the same image. To support some use cases (like Consul Connect
sidecars) we need to support the use of different docker images.