Add AppCreation object to handle creation process
The addition of the optional Graylog stream is transparent to
the App if GRAYLOG_ENABLED is not present.

Create

A conditional validation is present to avoid creating apps without
streams. If the checkbox is ticked the Stream creation has to be
successful in order for the App to be created.
An invalid App is not processed.