Conditionally render the checkbox
The checkbox is disabled by default on creation and is rendered conditionally
based on the GRAYLOG_ENABLED ENV var and presence of the associated
Graylog Stream.

If an application id edited and there is no stream, the option to add it
is visible in the form. Alternatively the stream is automatically updated
based on the presence of the repository_name field.