Add AppUpdate object to handle update process
Update

If an App has a stream and the integration is enabled the
index_set_id is updated accordingly.

If an App doesn't have a stream, the checkbox to
add one will be present in the UI. A JSON request will add a stream if
any valid graylog parameter is present.

The process is skipped if the App is invalid