Add rake task to associate current apps
In order to properly integrate an existing Rocksteady environment,
add a rake task that will associate existing apps with related
Graylog streams. The task is idempotent and ignores apps with existing
graylog_streams.

If a stream cannot be matched due to naming inconsistencies it will not
be associated and a warning bill be logged.