Graylog: Handle stream shares on Graylog 4
Infrastructure recently did a major version upgrade to our
centralized logging service, Graylog (8th Feb for staging and
Feb 24th for prod). This update changed the way permissions to
access streams work.

There are changes in how Roles are managed now:

    Additionally, since Graylog 4.0 now supports “sharing”
    functionality, granting access to streams and dashboards is
    no longer part of the "edit Roles" capability.

From the website:

    https://docs.graylog.org/en/4.0/pages/permission_management.html#roles

So now roles just define which actions you can take, but not
which streams you have access to. This should be done by sharing
the stream with each user who needs access to it. This can be
done through the POST /shares/entities endpoint.

To share a stream using this endpoint we need a stream id and
the ids of the users who would get access to it. For this we
pull users from the users endpoint (GET /users) and select
those with an allowed email.

We also have to reject the users that already have access to
the streams, in this case any user with admin power like the
Graylog API bot. The bot has now admin power because it is
needed in order to access the users endpoint.