Graylog: Update stream controller implementation
Replace the old way of setting user permissions on a stream
(using roles) with the new Graylog 4 way (using shares).

When no user ids are returned to share the steam with,
we fail with a custom error, so that we don't overwrite
the stream's permissions with an empty value (if that
happened, nobody would be able to see the stream except
for admin users).