Graylog: Remove Graylog ON/OFF feature
Rocksteady was created as an open-source project shared by
Altmetric and BotsAndUs. However, the latter have evolved into a
different solution and don't use Rocksteady anymore. Hence we
don't need the feature to enable or disable Graylog anymore.

Removing the `GRAYLOG_ENABLED` environment variable and the
logic associated with it simplifies code, tests and testing on
staging and live. If someone wanted to create apps with no
Graylog stream they can still do it by just not ticking the
Graylog checkbox in the web interface or not sending
`with_stream` with a value of `'1'` when using Rocksteady's
JSON API.